AbyssAC/requirements_cpu_core.txt
# 渊协议v5.2 - CPU版依赖（仅核心功能）
# 环境前提：Python 3.8~3.11（写在README里）

# 基础工具依赖
pillow>=9.0.0,<11.0.0
typing-extensions>=4.5.0,<5.0.0
networkx>=2.8.4,<3.2.0

# 数据处理依赖
numpy>=1.21.0,<1.26.0
scikit-learn>=1.0.0,<1.3.0
regex>=2022.3.15,<2024.0.0
python-dotenv>=1.0.0,<2.0.0

# 日志与监控依赖
tqdm>=4.64.0,<5.0.0
colorlog>=6.7.0,<7.0.0